[Stage3D] partial support for s3tc compressed textures (not yet working)

+++ src/parsing/streams.h
@@ -125,7 +125,7 @@ public:
 
 	inline void seekg(unsigned int offset)
 	{
-		if (offset >= lastcodepos-code)
+		if (offset >= (unsigned int)(lastcodepos-code))
 			codepos = lastcodepos;
 		else
 			codepos = code+offset;
